split up registration signup tab logic to sub components
