+++ _locales/es/messages.json
@@ -1308,7 +1308,7 @@
     "allUsersAreRandomly...": {
         "message": "Tu ID de Session es la dirección única que las personas pueden usar para contactarte en Session. Por diseño, tu ID de Session es totalmente anónima y privada, sin vínculo con tu identidad real."
     },
-    "generateSessionID": {
+    "createSessionID": {
         "message": "Crear ID de Session"
     },
     "recoveryPhrase": {
@@ -1319,9 +1319,6 @@
     },
     "enterDisplayName": {
         "message": "Ingresa un nombre para mostrar"
-    },
-    "enterSessionIDHere": {
-        "message": "Ingresa tu ID de Session"
     },
     "continueYourSession": {
         "message": "Continúa tu Session"
