+++ _locales/pl/messages.json
@@ -1308,7 +1308,7 @@
     "allUsersAreRandomly...": {
         "message": "Twój identyfikator Session to unikalny adres, za pomocą którego można się z Tobą kontaktować w Sesji. Bez połączenia z twoją prawdziwą tożsamością, identyfikator Session jest z założenia całkowicie anonimowy i prywatny."
     },
-    "generateSessionID": {
+    "createSessionID": {
         "message": "Utwórz identyfikator Session"
     },
     "recoveryPhrase": {
@@ -1319,9 +1319,6 @@
     },
     "enterDisplayName": {
         "message": "Wprowadź wyświetlaną nazwe"
-    },
-    "enterSessionIDHere": {
-        "message": "Wpisz swój identyfikator Session"
     },
     "continueYourSession": {
         "message": "Kontynuuj swoją sesję"
