+++ _locales/pt_BR/messages.json
@@ -1308,7 +1308,7 @@
     "allUsersAreRandomly...": {
         "message": "Seu ID Session é o endereço exclusivo que as pessoas podem usar para entrar em contato com você no Session. Sem conexão com sua identidade real, seu ID Session é totalmente anônimo e privado por definição."
     },
-    "generateSessionID": {
+    "createSessionID": {
         "message": "Criar ID Session"
     },
     "recoveryPhrase": {
@@ -1319,9 +1319,6 @@
     },
     "enterDisplayName": {
         "message": "Digite um nome de exibição"
-    },
-    "enterSessionIDHere": {
-        "message": "Digite seu ID Session"
     },
     "continueYourSession": {
         "message": "Continuar com seu Session"
