+++ ts/components/session/registration/SignInTab.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export enum SignInMode {
+  Default,
+  UsingRecoveryPhrase,
+  LinkDevice,
+}
+
+export interface Props {
+  signInMode: SignInMode;
+}
+
+export const SignInTab = (props: Props) => {};
