+++ ts/session/utils/User.ts
@@ -70,15 +70,8 @@ export async function getUserED25519KeyPair(): Promise<HexKeyPair | undefined> {
   return undefined;
 }
 
-/**
- * Returns the public key of this current device as a STRING, or throws an error
- */
 export function isRestoringFromSeed(): boolean {
-  const ourNumber = window.textsecure.storage.user.isRestoringFromSeed();
-  if (!ourNumber) {
-    throw new Error('ourNumber is not set');
-  }
-  return ourNumber;
+  return window.textsecure.storage.user.isRestoringFromSeed();
 }
 
 export function setRestoringFromSeed(isRestoring: boolean) {
